refactor(keystone): Move bin startup into server::startup
The `keystone` binary was a single 2172-line module: a 350-line
`main`, three 100-220 line helpers, and four blocks of copy-paste.
None of it was reachable from unit tests since binaries can't be
imported.

Relocate all startup logic into `openstack_keystone::server::startup`
submodules (args, tracing_init, audit, bootstrap, router, opa, raft,
listeners, background, shutdown); the bin is now a 3-line shell.
Existing tests move beside the code they cover.

Collapse the duplication along the way:
- tracing: one `deps_targets()` builder replaces 3 near-identical
  `Targets` filter chains
- event hooks: `subscribe!` macro replaces 13 repeated calls
- audit: extract `load_or_generate_kek`
- opa: `forward_opa_output` over `AsyncRead` replaces the twin
  stdout/stderr pumps; addr/client/poll/supervisor split into helpers

No functional change. Listen-phase debug timings keep their original
`listen_phase_start` origin.
