+++ crates/keystone/src/server.rs
@@ -18,3 +18,4 @@ pub mod http_metrics;
 pub mod listener;
 pub mod proxy_headers;
 pub mod request_cache;
+pub mod startup;
